Move Arbitrary modifiers to dedicated module
